New structural approach to gsheets: gsheets.py centralizes function calls, gsheets_client.py connects to the service, create_gsheet.py generates a new Google Sheet, and update_gsheet.py populates the new Google Sheet with values. I'm not going to write a client manager for output or even enable updating without creating a spreadsheet right now, but I'm trying to structure everything so that would be relatively easy to do. Next, I'll go through update_gsheet.py and hook it up to a Layout, eliminating redundant classes / redesigning them as plug-ins for my custom classes in spreadsheets.py.
