+++ gsheets/create_gsheet.py
@@ -1,30 +1,22 @@
-import os
-import sys
-# from property_gsheets import repeat_cell_builder
-from gsheets import GoogleSheetsAPIClient, GoogleSheet, SPREADSHEET_TITLE, SHEET_ONE_TITLE
-from property_gsheets import PropertySpreadsheet
+class GoogleSheet:
+  """Initialize with Google Sheet schema and build Google Sheet."""
+  def __init__(self, client, spreadsheet_title, sheet_one_title):
+    # TODO: error handling to ensure client is connected and auth'd
+    self.client = client
+    self.sheet_body = {
+        "properties": {
+          "title": spreadsheet_title
+        },
+        "sheets": [
+          {
+            "properties": {
+              "title": sheet_one_title
+            }
+          }
+        ]
+      }
+    
+    # Create spreadsheet
+    self.spreadsheet = self.client.spreadsheets().create(body=self.sheet_body).execute()
 
-
-# Locate parent directory to import property data scripts
-current_dir = os.path.dirname(os.path.realpath(__file__))
-parent_dir = os.path.dirname(current_dir)
-sys.path.insert(0, parent_dir)
-
-from property_level_analysis import address_data_to_gsheet
-
-# DATETIME_STRING used to parse JSON filenames for addresses containing ADDRESS_STRING
-ADDRESS_STRING = "7236 S Bell St"
-DATETIME_STRING = "2025-10-28"
-
-
-def main():
-  address_dict = address_data_to_gsheet(ADDRESS_STRING, DATETIME_STRING)
-  client = GoogleSheetsAPIClient()
-  gsheet = GoogleSheet(client.client, SPREADSHEET_TITLE, SHEET_ONE_TITLE)
-  property_gsheet = PropertySpreadsheet(address_dict, gsheet)
-  property_gsheet.update_values()
-  property_gsheet.update_format()
-
-
-if __name__ == "__main__":
-  main()
+    print(f"Spreadsheet created: {self.spreadsheet.get("spreadsheetUrl")}")
